Use normal constructor but be very clear that we are only modifying the payload.
The original code made it look like profile and durability
requirements could be changed in the visitor, and it took me a moment
to realize that it wasn't.  The amended code ran the risk of
incompletely initialized read_ts and write_ts in future (although the
code itself was correct).  This will hopefully be obvious and not
present risks.